+++ hw-ledger-transports/browser.ts
@@ -1,9 +1,9 @@
 
-import type Transport from 'https://esm.sh/@ledgerhq/hw-transport@6.28.0';
+import type Transport from 'https://esm.sh/@ledgerhq/hw-transport@6.28.1';
 import type { TransportDef } from './types.ts';
 
-import LedgerWebHid from 'https://esm.sh/@ledgerhq/hw-transport-webhid@6.27.11';
-import LedgerWebUsb from 'https://esm.sh/@ledgerhq/hw-transport-webusb@6.27.11';
+import LedgerWebHid from 'https://esm.sh/@ledgerhq/hw-transport-webhid@6.27.12';
+import LedgerWebUsb from 'https://esm.sh/@ledgerhq/hw-transport-webusb@6.27.12';
 
 export { packageInfo } from './packageInfo.ts';