+++ hw-ledger-transports/node.ts
@@ -1,8 +1,8 @@
 
-import type Transport from 'https://esm.sh/@ledgerhq/hw-transport@6.28.0';
+import type Transport from 'https://esm.sh/@ledgerhq/hw-transport@6.28.1';
 import type { TransportDef } from './types.ts';
 
-import LedgerHid from 'https://esm.sh/@ledgerhq/hw-transport-node-hid-singleton@6.28.8';
+import LedgerHid from 'https://esm.sh/@ledgerhq/hw-transport-node-hid-singleton@6.28.9';
 
 export { packageInfo } from './packageInfo.ts';