+++ util-crypto/address/eq.ts
@@ -1,7 +1,7 @@
 
-import type { HexString } from 'https://deno.land/x/polkadot@0.2.27/util/types.ts';
+import type { HexString } from 'https://deno.land/x/polkadot/util/types.ts';
 
-import { u8aEq } from 'https://deno.land/x/polkadot@0.2.27/util/mod.ts';
+import { u8aEq } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { decodeAddress } from './decode.ts';
 
@@ -14,7 +14,7 @@ import { decodeAddress } from './decode.ts';
  * <BR>
  *
  * ```javascript
- * import { u8aEq } from 'https://deno.land/x/polkadot@0.2.27/util/mod.ts';
+ * import { u8aEq } from 'https://deno.land/x/polkadot/util/mod.ts';
  *
  * u8aEq(new Uint8Array([0x68, 0x65]), new Uint8Array([0x68, 0x65])); // true
  * ```