+++ util/is/jsonObject.ts
@@ -16,7 +16,7 @@ interface ObjectIndexed {
  * <BR>
  *
  * ```javascript
- * import { isJsonObject } from 'https://deno.land/x/polkadot@0.2.27/util/mod.ts';
+ * import { isJsonObject } from 'https://deno.land/x/polkadot/util/mod.ts';
  *
  * isJsonObject({}); // => true
  * isJsonObject({